treewide: Add multiple commits
- devreq_boost: Add llcc to boosting device

Mimmick Pixel 4 userspace DDR boost.

While at it, bump up app fork duration.

- mm: Boost when memory pressure becomes high

When memory pressure starts to balloon, allocation latency takes a hit
as the page allocator continuously tries to find a way to free up some
pages. Boosting when this occurs can significantly help reduce
allocation latency, preventing noticeable lag from being observed.

- kernel: Boost to the max for a short amount of time when zygote forks

In order to boost during the period from when a new app is forked to the
time that it is added to the top-app cpuset (which is a signal to
subsequently boost again), boost for a short amount of time when zygote
forks. Not all processes forked from zygote are top apps, hence the
decoupled boosting setup.

- drm/atomic: Boost CPU and DDR bus when committing a new frame

In order to reduce jank, request a CPU and DDR bus boost whenever a new
frame is ready to be rendered to the display. The boosts should be
sufficient enough to render 60 FPS without any dropped frames when there
is no significant external source of load.

Co-authored by: Pierre2324 <pdbbasketball@gmail.com>
Signed-off-by: Sultan Alsawaf <sultan@kerneltoast.com>
Signed-off-by: Pierre2324 <pdbbasketball@gmail.com>